requirements/requirements.txt
Django==2.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
Pillow==5.2.0
pytz==2018.5
